+++ windowlist@cobinja.de/files/windowlist@cobinja.de/4.0/applet.js
@@ -49,6 +49,8 @@ const DEFAULT_ICON_SIZE = 22;
 const MINIMUM_ICON_SIZE = 16;
 const ICON_HEIGHT_FACTOR = 0.8;
 
+const FLASH_INTERVAL = 500;
+
 const PANEL_EDIT_MODE_KEY = "panel-edit-mode";
 
 const CobiCaptionType = {
@@ -770,11 +772,12 @@ class CobiAppButton {
     this._applet = applet;
     this._app = app;
     this._settings = this._applet._settings;
+    this._needsAttention = false;
     this._signalManager = new SignalManager.SignalManager(null);
     
     this._pinned = false;
     
-    this.actor = new St.BoxLayout({style_class: "window-list-item-box",
+    this.actor = new St.BoxLayout({style_class: "grouped-window-list-item-box",
                                    track_hover: true,
                                    can_focus: true,
                                    reactive: true
@@ -818,8 +821,6 @@ class CobiAppButton {
     this._windows = [];
     this._currentWindow = null;
     
-    this.actor.add_style_pseudo_class("neutral");
-    
     this._updateOrientation();
     
     this._contextMenuManager = new PopupMenu.PopupMenuManager(this);
@@ -838,7 +839,6 @@ class CobiAppButton {
     this._signalManager.connect(this.actor, "notify::hover", this._updateVisualState, this);
     
     this._signalManager.connect(Main.themeManager, "theme-set", Lang.bind(this, function() {
-      this.actor.remove_style_pseudo_class("neutral");
       this.updateView();
     }), this);
     this._signalManager.connect(St.TextureCache.get_default(), "icon-theme-changed", this.updateIcon, this);
@@ -920,7 +920,7 @@ class CobiAppButton {
     this._signalManager.connect(metaWindow, "notify::wm-class", this._onWmClassChanged, this);
     this._signalManager.connect(metaWindow, "workspace-changed", this._onWindowWorkspaceChanged, this);
     
-    this.actor.remove_style_pseudo_class("neutral");
+    this.actor.add_style_pseudo_class("active");
     this._updateTooltip();
     if (this._windows.length == 1) {
       this._workspace.menuManager.addMenu(this.menu);
@@ -948,7 +948,6 @@ class CobiAppButton {
       if (!this._currentWindow) {
         this.actor.remove_style_pseudo_class("focus");
         this.actor.remove_style_pseudo_class("active");
-        this.actor.add_style_pseudo_class("neutral");
         this._workspace.menuManager.removeMenu(this.menu);
       }
     }
@@ -1117,21 +1116,8 @@ class CobiAppButton {
   }
   
   _updateVisualState() {
-    if (this._currentWindow) {
-      if (this.actor.has_style_pseudo_class("neutral")) {
-        this.actor.remove_style_pseudo_class("neutral");
-      }
-    }
-    else {
-      if (this.actor.has_style_pseudo_class("focus")) {
-        this.actor.remove_style_pseudo_class("focus");
-      }
-      if (!this.actor.has_style_pseudo_class("neutral")) {
-        this.actor.add_style_pseudo_class("neutral");
-      }
-      else if (this.actor.has_style_pseudo_class("hover")) {
-        this.actor.remove_style_pseudo_class("neutral");
-      }
+    if (!this._currentWindow) {
+      this.actor.remove_style_pseudo_class("focus");
     }
   }
   
@@ -1164,17 +1150,45 @@ class CobiAppButton {
     }
   }
   
+  _flashButton() {
+    if (!this._needsAttention){
+      return;
+    }
+
+    let counter = 0;
+    let sc = "grouped-window-list-item-demands-attention";
+    
+    global.log("Start flashing");
+    this.actor.add_style_class_name(sc);
+
+    Mainloop.timeout_add(FLASH_INTERVAL, () => {
+      if (!this._needsAttention) {
+        return false;
+      }
+      if (this.actor.has_style_class_name(sc)) {
+        global.log("Flash off");
+        this.actor.remove_style_class_name(sc);
+      }
+      else {
+        global.log("Flash on");
+        this.actor.add_style_class_name(sc);
+      }
+      let result = counter <= 4;
+      counter++;
+      return result;
+    });
+  }
+  
   _updateUrgentState() {
+    if (this._needsAttention) {
+      return;
+    }
+    
     let state = this._windows.some(function(win) {
       return win.urgent || win.demands_attention;
     });
-    
-    if (state) {
-      this.actor.add_style_class_name("window-list-item-demands-attention");
-    }
-    else {
-      this.actor.remove_style_class_name("window-list-item-demands-attention");
-    }
+    this._needsAttention = state;
+    this._flashButton();
   }
   
   _updateFocus() {
@@ -1182,6 +1196,7 @@ class CobiAppButton {
       let metaWindow = this._windows[i];
       if (hasFocus(metaWindow) && !metaWindow.minimized) {
         this.actor.add_style_pseudo_class("focus");
+        this.actor.remove_style_class_name("grouped-window-list-item-demands-attention");
         this._currentWindow = metaWindow;
         this._updateLabel();
         break;